+++ mail/mail-common/src/user_context/events/messages.rs
@@ -1,5 +1,5 @@
 use crate::events::MessageEvent;
-use crate::models::Message;
+use crate::models::{ConversationLabel, Message};
 use crate::AppError;
 use proton_core_common::events::Action;
 use stash::params;
@@ -21,6 +21,11 @@ pub async fn handle_message_events(
             }
             Action::Create => {
                 if let Some(message) = &message_event.message {
+                    ConversationLabel::create_or_update_from_message_metadata(
+                        &[message.clone()],
+                        tx,
+                    )
+                    .await?;
                     Message::create_or_update_messages_from_metadata(vec![message.clone()], tx)
                         .await?;
                 } else {
@@ -29,6 +34,11 @@ pub async fn handle_message_events(
             }
             Action::Update | Action::UpdateFlags => {
                 if let Some(message) = &message_event.message {
+                    ConversationLabel::create_or_update_from_message_metadata(
+                        &[message.clone()],
+                        tx,
+                    )
+                    .await?;
                     Message::create_or_update_messages_from_metadata(vec![message.clone()], tx)
                         .await?;
                 } else {